+++ src/settings/MappingProfiles/MappingProfiles.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const MappingProfiles = () => {
+  return (
+    <span>Mapping profiles</span>
+  );
+};
+
+export default MappingProfiles;