+++ test/bigtest/interactors/settings/index.js
@@ -1,0 +1,1 @@
+export { default as SettingsSectionsPaneInteractor } from './SettingsSectionsPaneInteractor';